% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qryResults.R
\name{qryResults}
\alias{qryResults}
\title{A SQL query function, preconnected to SPREPORTING}
\usage{
qryResults(sql)
}
\arguments{
\item{sql}{SQL code to run}
}
\description{
This function runs SQL on SPREPORTING.
}
\keyword{SQL}

